refactor: S29.05 extract shared semihosting block I/O from diskio
Lift the BKPT 0xAB trap, the open/read/write/seek/length/close primitives,
the 8 MiB FAT16 geometry, and the open-or-create-sparse image logic out of
diskio.c into a new Bdd/Targets/Common/SemihostingDisk TU. diskio.c now maps
the ChaN-FatFs disk_* contract onto SemihostingDisk_{EnsureReady,IsReady,
Read,Write}, preserving its RES_NOTRDY / RES_PARERR / RES_ERROR distinctions.
This lets the upcoming FreeRTOS-Plus-FAT FF_Disk_t driver share byte-for-byte
the same host image and trap. No behaviour change. Verified both FreeRTOS
cross targets still build.
